GifCaptcha Draft 3
able to generate random number 0-7;
-------------------------------------
still a bit confused about how to plug it into the solution. Need to insert dog div in a random part of the mainContentWrapper.children array. Was thinking about using splice but to my understanding, it adds a new item to an array at any given point...